Fix style usage, simplify nav, fix router on finish
Factored out some styles where it was possible. Changed the navigation
buttons to just use the ternary operator to choose a container instead
of wrapping the whole thing in a container.

Previously, router upon "finish" would push the askMenstruate page, and
then after that was closed, simply return to the last page of
onboarding. This dismisses the onboarding page first and then pushes the
askMenstruate page, so it will return to the page before onboarding.

TODO: Currently the back/next buttons don't scroll to the top of the
screen on pressed, so the user could miss info at the top.